UI changes, logic fixes for timer, adding distinct player names
