+++ server/src/rooms/experimentRoom.ts
@@ -3,6 +3,9 @@ import DataLogger from '../dataLogger';
 import {Player, Zone, RoomState, Phase} from './schema/experimentSchema';
 import {config} from '../config';
 import { randomInt } from 'node:crypto';
+import {generateDistinctColor, getContrastingTextColor} from '../utils';
+import { faker } from '@faker-js/faker';
+import { kMaxLength } from 'node:buffer';
 
 export class ExperimentRoom extends Room<RoomState> {
 	state = new RoomState();
@@ -35,11 +38,17 @@ export class ExperimentRoom extends Room<RoomState> {
 		const player = new Player();
 		player.x = config.player.startX; // Start at center
 		player.y = config.player.startY;
-		player.name = options.name || `Player ${client.sessionId.slice(0, 4)}`;		
+		// player.name = options.name || `Player ${client.sessionId.slice(0, 4)}`;
+		const adjective = faker.word.adjective({ length: { min: 5, max: 8 }, strategy: "closest" });
+		const noun = faker.word.noun({ length: { min: 5, max: 8 }, strategy: "closest" });
+		// Capitalize first letter of each word
+		const capitalizedAdjective = adjective.charAt(0).toUpperCase() + adjective.slice(1);
+		const capitalizedNoun = noun.charAt(0).toUpperCase() + noun.slice(1);
+		player.name = `${capitalizedAdjective}${capitalizedNoun}`;		
 
 		// Combine zone and player hues to ensure player colors are distinct from both zones and other players
 		const allUsedHues = new Set([...this.zoneHues, ...this.playerHues]);
-		player.color = this.generateDistinctColor(allUsedHues, config.player.minHueDifference);
+		player.color = generateDistinctColor(allUsedHues, config.player.minHueDifference);
 
 		// The new hue was added to allUsedHues, so we need to find it and add to playerHues
 		// by comparing the sets
@@ -51,7 +60,7 @@ export class ExperimentRoom extends Room<RoomState> {
 		}
 
 		// Generate contrasting text color for readability
-		player.textColor = this.getContrastingTextColor(player.color);
+		player.textColor = getContrastingTextColor(player.color);
 		player.emote = '';
 
 		this.state.players.set(client.sessionId, player);
@@ -78,109 +87,20 @@ export class ExperimentRoom extends Room<RoomState> {
 		this.state.players.delete(client.sessionId);
 	}
 
-	// Helper method to generate distinct colors using HSL color space
-	// This ensures colors are dramatically different in hue
-	private generateDistinctColor(usedHues: Set<number>, minHueDifference: number = 60): string {
-		const hueOptions: number[] = [];
-
-		// Generate potential hues with good separation
-		for (let hue = 0; hue < 360; hue += minHueDifference) {
-			let isDistinct = true;
-			for (const usedHue of usedHues) {
-				const hueDiff = Math.min(
-					Math.abs(hue - usedHue),
-					360 - Math.abs(hue - usedHue)
-				);
-				if (hueDiff < minHueDifference) {
-					isDistinct = false;
-					break;
-				}
+	// Helper method to check player zone
+	private checkZone(player_x: number, player_y: number, player_radius?: number): number {
+		let zone_id: number = -1;
+		if (!player_radius) {
+			player_radius = 0;
+		}
+		this.state.zones.forEach((zone) => {
+			if (Math.hypot(Math.abs(player_x - zone.x), Math.abs(player_y - zone.y)) <= (zone.radius + player_radius)) {
+				zone_id = zone.id;
 			}
-			if (isDistinct) {
-				hueOptions.push(hue);
-			}
-		}
+		});
 
-		// If no distinct hues available, reduce the minimum difference
-		if (hueOptions.length === 0 && minHueDifference > config.player.minHueDifferenceFallback) {
-			return this.generateDistinctColor(usedHues, minHueDifference - 10);
-		}
-
-		// Select random hue from available options
-		const selectedHue = hueOptions.length > 0
-			? hueOptions[Math.floor(Math.random() * hueOptions.length)]
-			: Math.floor(Math.random() * 360);
-
-		usedHues.add(selectedHue);
-
-		// Use high saturation and medium lightness for vibrant colors
-		const saturation = config.colors.saturation.min + Math.random() * (config.colors.saturation.max - config.colors.saturation.min);
-		const lightness = config.colors.lightness.min + Math.random() * (config.colors.lightness.max - config.colors.lightness.min);
-
-		return this.hslToHex(selectedHue, saturation, lightness);
+		return zone_id;
 	}
-
-	// Helper method to convert HSL to hex color
-	private hslToHex(h: number, s: number, l: number): string {
-		s /= 100;
-		l /= 100;
-
-		const c = (1 - Math.abs(2 * l - 1)) * s;
-		const x = c * (1 - Math.abs((h / 60) % 2 - 1));
-		const m = l - c / 2;
-
-		let r = 0, g = 0, b = 0;
-
-		if (h >= 0 && h < 60) {
-			r = c; g = x; b = 0;
-		} else if (h >= 60 && h < 120) {
-			r = x; g = c; b = 0;
-		} else if (h >= 120 && h < 180) {
-			r = 0; g = c; b = x;
-		} else if (h >= 180 && h < 240) {
-			r = 0; g = x; b = c;
-		} else if (h >= 240 && h < 300) {
-			r = x; g = 0; b = c;
-		} else if (h >= 300 && h < 360) {
-			r = c; g = 0; b = x;
-		}
-
-		const toHex = (value: number) => {
-			const hex = Math.round((value + m) * 255).toString(16);
-			return hex.length === 1 ? '0' + hex : hex;
-		};
-
-		return `#${toHex(r)}${toHex(g)}${toHex(b)}`.toUpperCase();
-	}
-
-	// Helper method to calculate contrasting text color (black or white)
-	private getContrastingTextColor(hexColor: string): string {
-		// Convert hex to RGB
-		const r = parseInt(hexColor.slice(1, 3), 16);
-		const g = parseInt(hexColor.slice(3, 5), 16);
-		const b = parseInt(hexColor.slice(5, 7), 16);
-
-		// Calculate relative luminance using WCAG formula
-		const luminance = (0.299 * r + 0.587 * g + 0.114 * b) / 255;
-
-		// Return black for light backgrounds, white for dark backgrounds
-		return luminance > 0.5 ? '#000000' : '#FFFFFF';
-	}
-
-    // Helper method to check player zone
-    private checkZone(player_x: number, player_y: number, player_radius?: number): number {
-        let zone_id: number = -1;
-        if (!player_radius) {
-            player_radius = 0;
-        }
-        this.state.zones.forEach((zone) => {
-            if (Math.hypot(Math.abs(player_x - zone.x), Math.abs(player_y - zone.y)) <= (zone.radius + player_radius)) {
-                zone_id = zone.id;
-            }
-        });
-
-        return zone_id;
-    }
 
     // Helper method to create zone
     private makeZone(id: number, x: number, y: number, radius: number): Zone {
@@ -191,7 +111,7 @@ export class ExperimentRoom extends Room<RoomState> {
 		zone.y = y;
 		zone.radius = radius;
 		// Generate distinct color for zone with dramatic hue separation
-		zone.color = this.generateDistinctColor(this.zoneHues, config.zones.minHueDifference);
+		zone.color = generateDistinctColor(this.zoneHues, config.zones.minHueDifference);
 
         return zone;
     }
@@ -365,6 +285,15 @@ export class ExperimentRoom extends Room<RoomState> {
 			player.x += data.x;
 			player.y += data.y;
 			console.log(client.sessionId + ' at, x: ' + player.x, 'y: ' + player.y);
+
+			// Clamp player position to world bounds (accounting for player radius)
+			const minX = player.radius;
+			const maxX = config.world.width - player.radius;
+			const minY = player.radius;
+			const maxY = config.world.height - player.radius;
+
+			player.x = Math.max(minX, Math.min(maxX, player.x));
+			player.y = Math.max(minY, Math.min(maxY, player.y));
 
             player.zone = this.checkZone(player.x, player.y, player.radius);
 		});
